+++ src/app/models/suite.ts
@@ -7,11 +7,11 @@ export class Suite {
   created_at: string;
   updated_at: string;
   statistic: Statistic;
-  path = './suite';
+  path = 'suite';
   constructor (suite) {
     if (suite == null) {
       suite = this.create_default_suite();
-    } else if (suite.path === './run') {
+    } else if (suite.path === 'run') {
       suite = this.create_suite_by_run(suite);
     }
     this.id = suite['id'];
@@ -22,13 +22,13 @@ export class Suite {
     this.statistic = new Statistic({0: suite['statistic'][0]['count']});
   }
   create_default_suite() {
-    return {'id': 'id_loading', 'name': 'name_loading',
+    return {'id': 'id_loading', 'name': 'name_loading', 'path': 'suite',
       'product_id': 'product_id_loading',
       'created_at': 'created_at_loading', 'updated_at': 'updated_at_loading',
       'statistic': {0: 0}};
   }
   create_suite_by_run(suite) {
-    return {'id': suite['id'], 'name': suite['name'],
+    return {'id': suite['id'], 'name': suite['name'], 'path': 'run',
       'product_id': 0,
       'created_at': suite['created_at'].split(' +')[0], 'updated_at': suite['updated_at'].split(' +')[0],
       'statistic': {0: 0}};