+++ Changelog.md
@@ -2,6 +2,7 @@
 ## master (unreleased)
 ## Added
     * New componens: status filter. It is status with count of elements
+    * Return default untested status color to transparent
 
 ## [0.5.8] - 03.11.2019
 ## Fixed