+++ Changelog.md
@@ -4,6 +4,7 @@
     * New componens: status filter. It is status with count of elements
     * Return default untested status color to transparent
     * New settings button to plans (need to hover item for see it)
+    * Add new runs and suites list (redisign for using material)
 
 ## [0.5.8] - 03.11.2019
 ## Fixed